+++ src/main/java/org/sitmun/authorization/client/service/Profile.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import org.sitmun.domain.application.Application;
-import org.sitmun.domain.background.Background;
 import org.sitmun.domain.cartography.Cartography;
 import org.sitmun.domain.cartography.permission.CartographyPermission;
 import org.sitmun.domain.configuration.ConfigurationParameter;
@@ -22,7 +21,7 @@ import org.sitmun.domain.tree.node.TreeNode;
 public class Profile {
   private Application application;
   private Territory territory;
-  private List<Background> backgrounds;
+  private List<ApplicationBackgroundView> backgrounds;
   private List<CartographyPermission> groups;
   private List<Service> services;
   private List<Cartography> layers;